Level Mesh Fixes & Texture Compression
I compressed the textures down to 128x and fixed the UVs, Normals, etc... of the level mesh parts. All level parts have assigned materials and the UVs are set against the door sizes to keep consistency. More or less it looks really good. But it needs interactability and sfx to really pull it together